fix(mobile): moved type and constants for buttons, update wheel bb-571
