--- apps/mobile/src/screens/chat/libs/components/chat-buttons/libs/components/explain-accept-task-buttons/constants/ai-response-text.constant.ts
@@ -1,3 +1,0 @@
-const AI_RESPONSE = "Way to go! I will check on you!";
-
-export { AI_RESPONSE };
