--- apps/mobile/src/screens/chat/libs/components/chat-buttons/libs/components/explain-accept-task-buttons/constants/constants.ts
@@ -1,2 +1,0 @@
-export { AI_RESPONSE } from "./ai-response-text.constant";
-export { EXPLAIN_TEXT } from "./explain-text.constant";
