--- apps/mobile/src/screens/chat/libs/components/chat-buttons/libs/components/explain-accept-task-buttons/constants/explain-text.constant.ts
@@ -1,3 +1,0 @@
-const EXPLAIN_TEXT = "Let me know how it looks for you";
-
-export { EXPLAIN_TEXT };
