--- apps/mobile/src/screens/chat/libs/components/chat-buttons/libs/components/explain-accept-task-buttons/types/root-stack-parameter-list.type.ts
@@ -1,7 +1,0 @@
-import { type BottomTabScreenName } from "~/libs/enums/enums";
-
-type RootStackParameterList = {
-	[BottomTabScreenName.TASKS]: undefined;
-};
-
-export { type RootStackParameterList };
