--- apps/mobile/src/screens/chat/libs/components/chat-buttons/libs/components/explain-accept-task-buttons/types/types.ts
@@ -1,1 +1,0 @@
-export { type RootStackParameterList } from "./root-stack-parameter-list.type";
