+++ apps/mobile/src/screens/chat/libs/components/chat-buttons/libs/constants/ai-response-text.constant.ts
