+++ apps/mobile/src/screens/chat/libs/components/chat-buttons/libs/constants/explain-text.constant.ts
